added RideRequest & RidePolicy

+++ app/Http/Controllers/RideController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Ride;
-use Illuminate\Http\Request;
+use App\Http\Requests\RideRequest;
 
 class RideController extends Controller
 {
@@ -14,7 +14,6 @@ class RideController extends Controller
 
     public function index()
     {
-        //$rides = Ride::all();
         $rides = auth()->user()->rides;
         return view('rides.index', compact('rides'));
     }
@@ -24,56 +23,45 @@ class RideController extends Controller
         return view('rides.create');
     }
 
-    public function store(Request $request)
+    public function store(RideRequest $request)
     {
-        $attributes = request()->validate([
-            'title'           => ['required', 'min:3'],
-            'date'            => ['required', 'date'],
-            'distance'        => 'required',
-            'duration'        => 'required',
-            'type'            => 'required',
-            'weather'         => 'nullable',
-            'link_strava'     => 'nullable',
-            'link_garmin'     => 'nullable',
-            'remarks'         => 'nullable'
-        ]);
-
-        $attributes['user_id'] = auth()->id();
-
-        Ride::create($attributes);
+        Ride::create(
+            ['user_id' => auth()->id()] +
+            request(['title', 'date', 'distance', 'duration', 'type', 'weather', 'link_strava', 'link_garmin', 'remarks'])
+        );
 
         return redirect('/rides');
     }
 
     public function show(Ride $ride)
     {
+        $this->authorize('update', $ride);
+
         return view('rides.show', compact('ride'));
     }
 
     public function edit(Ride $ride)
     {
-        return view('rides.edit', compact('ride'));
+       $this->authorize('update', $ride);
+
+       return view('rides.edit', compact('ride'));
     }
 
-    public function update(Request $request, Ride $ride)
+    public function update(RideRequest $request, Ride $ride)
     {
-        $ride->update(request()->validate([
-            'title'       => ['required', 'min:3'],
-            'date'        => ['required', 'date'],
-            'distance'    => 'required',
-            'duration'    => 'required',
-            'type'        => 'required',
-            'weather'     => 'nullable',
-            'link_strava' => 'nullable',
-            'link_garmin' => 'nullable',
-            'remarks'     => 'nullable'
-        ]));
+        $this->authorize('update', $ride);
+
+        $ride->update(
+            request(['title', 'date', 'distance', 'duration', 'type', 'weather', 'link_strava', 'link_garmin', 'remarks'])
+        );
 
         return redirect('/rides/' . $ride->id);
     }
 
     public function destroy(Ride $ride)
     {
+        $this->authorize('update', $ride);
+
         $ride->delete();
 
         return redirect('/rides');
